server_agent/requirements.txt
livekit>=1.0
livekit-agents[google,deepgram,elevenlabs,silero,turn_detector,assemblyai]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
fastapi
uvicorn[standard]
pydantic